openssh_cert: drop signature_algorithm choices
They have in the past grown outdated and prevented application of more
recent algorithms (which have been added in the past few years).
Maintaining a list of choiced would require continuous monitoring in
order to provide the current up to date and prevent usage of new
algorithms if new choices are not added to the code base in time.